// backend/src/PremierLeague/types/API-Football/teams.ts
import { APIResponse } from "./response.js";

export interface APITeam {
    team: {
        id: number;
        name: string;
        code: string;
        country: string;
        founded: number;
        national: boolean;
        logo: string
    };
    venue: {
        id: number;
        name: string;
        address: string;
        city: string;
        capacity: number;
        surface: string;
        image: string;
    };
}

export type TeamsAPIResponse = APIResponse<APITeam>